requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
openpyxl
xlrd
plotly
markdown
tabulate